skynetsys/pyproject.toml
[tool.poetry]
name = "skynetsys"
version = "0.1.0"
description = "Evan's Skynet server"
authors = ["Evan Chen <evan@evanchen.cc>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"
psutil = "^5.8.0"

[tool.poetry.dev-dependencies]
jedi = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
